Separated the deploy_stack classes into multiple files and fixed some issues with the example files
