+++ .gitignore
@@ -6,3 +6,4 @@ dist/**
 .idea
 .vscode
 .DS_Store
+**/node_modules
