+++ examples/cloudformation/my-project.cf.yaml
