+++ examples/cloudformation/support/bucket-policy.cf.yaml
