+++ examples/cloudformation/support/s3-watchdog-lambda.cf.yaml
