+++ examples/src/kmsParameters/package.json
